:bug: Reference guestID from ConfigInfo when evaluating policies
Currently, the policy evaluation code looks at the VM.Config.Guest to
get the guest info (OS family and ID) if the VM exists.  The problem
is that this only works if the VM is powered on and VM tools is
running.

This change changes this behavior to reference the ConfigInfo.guestID
field instead.  GuestOSFamily is derived from the ID.