Recreate SPU for snapshot if it has a wrong extension name (#1365)
This bug could happen when upgrading from 9.0.x to 9.1. Since 9.0.x contains code that creates SPU for snapshot with wrong extension name if snapshot is enabled.
That wasn't fixed under the impression that snapshot feature gate will be disabled for 9.0.x. But when upgrading, Feature gate will be flipped first then there is a tiny window that old VMOP could create SPU with wrong extensionName. The extensionName is immutable. Then VMOP can't correct it.