Account for swap file in StorageQuota validation (#1360)
When creating a new VM, the storage quota webhook should account
calculate the potential swap file size and add this to the calculation
for requested capacity. This calculation is the difference between
total memory and reserved memory.

Additionally this PR adds additional context to some of the errors
returned by the quota webhook and prefixes these error messages with
the webhook path being called. There has been some ambiguity in the
past as to the source of the errors. So this addition is to help
clarify if the source of the error is the VMOp extension or the calling
CSI webhook.