Preserve existing ConfigSpec TagSpecs during placement
When creating a VM, tag specs from vSphere polices may be
applied to the ConfigSpec but when applying tag specs for
affinity/anti-affinity, the existing tag specs would be
removed. This would only happen if the VM had non-filtered
out labels.

Remove the existing sorting of the TagSpecs in placement
ConfigSpec: order does not matter for this.

Use ConsistOf instead of ContainElements in a few other
places for stronger assertions.