Handle empty image kind ResolveImageName()
For a VM that was created when like v1a1 or v1a2 was the
stable version, the VM Spec.Image.Name will get backfilled
during version conversion from the Spec.ImageName but that
will leave the Kind unset.

We do have some code to handle backfilling Kind during just
the VM the intila VC VM create reconcile but not in the usual
reconcile. Keep this change smaller and defer doing that.

There is a thought that the Spec.Image could be a part of
the schema upgrade, but it doesn't fit well in that. The
image may not exist presently but it may be (re)created
later. We couldn't really set the feature bit until the
image kind is resolved, but in the meanwhile, the webhook
think the schema isn't upgrade. I don't want to add like
optional activated features right now.