Use PlaceVmsXCluster for most placement decisions
Previously we would use PlaceVm if there was a single candidate, but
now use the more modern & preferred PlaceVmsXCluster. The only time
we will use old PlaceVm is for instance storage because PlaceVmsXCluster
does not support the magic disk ID.

Also fix the incorrect assumption that instance storage can not be
used on stretched setups.

Clean up some old placement related cruft: the "ZonePlacement" has
been just true for sometime so treat it as implied and remove the
host placement bool since that is only used for Instance Storage.