Return early after adding instance storage volumes
Instance storage does not work with fast deploy so we disable the
feature in the VM scope. However, the instance storage volumes are
copied from the class to the VM Spec during the "first" reconcile
so the feature would not be disabled. Return early after adding the
volumes to the controller will observe the updated VM and disable
the feature.

This was an existing issue that was masked by the previous behavior
of using PlaceVM when there was only a single candidate. In that
case, with fast deploy still enabled the selected zone would not
be set in the labels but the VM Spec would have the instance storage
volumes added. On the next reconcile, fast deploy would be disabled,
but we would not have the zone set so we would think that we did
not need to do instance storage placement to select the host but
zone placement to pick the zone.

We could tweak the behavior of doesVMNeedPlacement() instead here
but my longer term is to get rid of that short circuit since for
fast deploy we'll always need the datastores, so let's just make
sure the fast deploy is always disabled for the reconcile of an
instance storage VM.