:sparkles: Add Typos check to prevent typos (#1395)
This change introduces a new make target and a new Github Workflow
step to find any typos in our code.

Also fix all the typos present as of now.

I have tried to minimize the exclusion rules to bare minimum.  It is
impossible to fix all of those since some of them are a result of
typos / errors in upstream code (e.g., ANDed).