:bug: Do not attempt to encrypt VMs erroneously (#1396)
We have a bug where if a native key provider is configured, VM
operator will attempt to encrypt that VM.  This is incorrect.  The
only scenarios where a VM should be encrypted is:
- if it specifies an encryption storage class
- uses a vTPM (existing, or new device being added)
(and a native key provider, or a custom one is specified via BYOK).

This results in the VM reporting the following Condition erroneously:
```
k get vm -n parunesh-ns parunesh-vm -o=json | jq -r '.status.conditions[] | select(.type == "VirtualMachineEncryptionSynced")'
{
  "lastTransitionTime": "2025-12-15T21:31:37Z",
  "message": "Must use encryption storage class or have vTPM when encrypting vm",
  "reason": "InvalidState",
  "status": "False",
  "type": "VirtualMachineEncryptionSynced"
}
```

This change fixes this bug.