:bug: Do not send in profile in CopyVirtualDisk API
CopyVirtualDisk ignores the Crypto and the Profile parameters. Crypto
was handled (not being sent). But, we were sending in the profile
parameter. The CopyVirtualDisk API is being fixed to handle both these
parameters correctly. Once that happens, we can't send in a profile
that is encryptinon enabled without sending in Crypto. That fails the
VM deployment.

This change omits the VM profile from the CopyVirtualDisk API call.