:bug: Do not reference VM class to get memory required for snapshot
In order to figure out the storage required from the StorageClass to
take the snapshot, we look at the VM class to get the memory required
for the snapshot in case of a memory-full snapshot. Imported VMs are
classless VMs, so that leads to snapshots being rejected for such VMs.

This change updates this workflow to reference the VM's
status.hardware to get the memory required for this VM's snapshot.