Skip SCSI hot-add validation before schema upgrade
This PR fixes a bug where if the VM is trying to attach volumes to explicit
non-zero SCSI controllers, and the VM image has a disk. Then, VM operator
fails to power on the VM because the mutation hook will try to add a new
controller at SCSI:0 for the bootdisk and the hot-add validation will reject that change.