Clean up VirtualMachine controller finish logging
If VM reconcile returned a NoRequeueError, include the reason in the
finished log since that is a very helpful indication as what the
reconcile did, and what it is expecting to kick the next reconcile.

While here, remove the old status changed and IP set logging. This was
done long ago due to the otherwise limited abilities of VMC LINT
which doesn't matter anymore, and don't seem to be generally very
helpful (especially the status updated one). We have a condition for
the VM's IP.