Support for non-removable volumes
This patch introduces support for marking spec.volumes entries in
a VM object as non-removable. This is to help prevent accidental
removal of the disks that originate from a VM image, such as the
disk(s) required to boot the VM. A user may mark a disk as removable
at any time, but by default a VM image's disks are treated as
non-removable.