Separate VM schema/object upgrade
This patch changes the way a VM is schema upgraded. Previously a VM
would be schema upgraded in the same patch as its object upgrade, i.e.
backfilling data. This change ensures the VM is schema upgraded in a
distinct K8s patch prior to modifying the VM's actual object data.
This change is to better handle K8s upgrades and conversion issues.